refactor(markdown): share the projection serializer from the engine
Move the projection-plan JSON serializer out of nteract-markdown-wasm and into
nteract-markdown-engine as a render_json module, and split a parse-free
project_from_mdast entry off project_markdown_with_options. The serializer was
portable std Rust over the public plan types with no wasm-specific code, so the
engine can own it and any crate that holds a MarkdownPlan can render the same
JSON. The engine label ("rust-wasm") is now a render_plan_json parameter so a
non-wasm caller can identify itself without changing the schema. The wasm crate
drops to an FFI shim that calls the engine and passes "rust-wasm", so the .md
projection JSON is byte-identical.

The Wasm-prefixed serializer structs become Json-prefixed now that they live in
the engine. The serializer tests move with the code.
